Moved activities into an activities package
