+++ app/src/androidTest/java/com/example/pantrypal/LoginTest.java
@@ -16,6 +16,9 @@ import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.filters.LargeTest;
 
+import com.example.pantrypal.activities.Login;
+import com.example.pantrypal.activities.MainActivity;
+
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
