+++ app/src/main/AndroidManifest.xml
@@ -15,10 +15,10 @@
         android:theme="@style/Theme.PantryPal"
         tools:targetApi="31">
         <activity
-            android:name=".Registration"
+            android:name=".activities.Registration"
             android:exported="false" />
         <activity
-            android:name=".Login"
+            android:name=".activities.Login"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -27,33 +27,33 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.MainActivity"
             android:exported="false" />
         <activity
             android:name=".NewRecipe"
             android:exported="false" />
         <activity
-            android:name=".Search"
+            android:name=".activities.Search"
             android:exported="false"
             android:theme="@style/Theme.PantryPal2"/>
         <activity
             android:name=".Favorites"
             android:exported="false" />
         <activity
-            android:name=".IngredientsSearch"
+            android:name=".activities.IngredientsSearch"
             android:exported="false"/>
         <activity
-            android:name=".ViewRecipe"
+            android:name=".activities.ViewRecipe"
             android:exported="false" />
-        <activity android:name="com.example.pantrypal.ViewMealDBRecipe" />
+        <activity android:name="com.example.pantrypal.activities.ViewMealDBRecipe" />
         <activity
-            android:name=".Home"
-            android:exported="false" />
-        <activity
-            android:name=".AddTags"
+            android:name=".activities.Home"
             android:exported="false" />
         <activity
-            android:name=".Reviews"
+            android:name=".activities.AddTags"
+            android:exported="false" />
+        <activity
+            android:name=".activities.Reviews"
             android:exported="false" />
 
     </application>
