+++ app/src/main/java/com/example/pantrypal/ResultsRecipe.java
@@ -8,36 +8,36 @@ public class ResultsRecipe {
     private String id;
     private List<String> ingredients;
 
-    ResultsRecipe() {
+    public ResultsRecipe() {
         this.title = "Title";
         this.id = "id String";
         this.ingredients = new ArrayList<>();
     }
 
-    ResultsRecipe(String title, String id) {
+    public ResultsRecipe(String title, String id) {
         this.title = title;
         this.id = id;
     }
 
-    ResultsRecipe(String title, String id, List<String> ingredients) {
+    public ResultsRecipe(String title, String id, List<String> ingredients) {
         this.title = title;
         this.id = id;
         this.ingredients = new ArrayList<>(ingredients);
     }
 
-    void setTitle(String title) {
+    public void setTitle(String title) {
         this.title = title;
     }
 
-    String getTitle() {
+    public String getTitle() {
         return title;
     }
 
-    void setId(String id) {
+    public void setId(String id) {
         this.id = id;
     }
 
-    String getId() {
+    public String getId() {
         return id;
     }
