+++ app/src/main/java/com/example/pantrypal/activities/AddTags.java
@@ -1,4 +1,4 @@
-package com.example.pantrypal;
+package com.example.pantrypal.activities;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -7,6 +7,9 @@ import android.widget.Button;
 import android.widget.CheckBox;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.pantrypal.NewRecipe;
+import com.example.pantrypal.R;
 
 import java.util.ArrayList;
 import java.util.Arrays;
