+++ app/src/main/java/com/example/pantrypal/activities/RecipeCard.java
@@ -1,4 +1,4 @@
-package com.example.pantrypal;
+package com.example.pantrypal.activities;
 
 import android.app.Activity;
 import android.content.Context;
