+++ app/src/main/java/com/example/pantrypal/activities/ViewMealDBRecipe.java
@@ -1,4 +1,4 @@
-package com.example.pantrypal;
+package com.example.pantrypal.activities;
 
 import android.content.Intent;
 import android.graphics.Bitmap;
@@ -11,6 +11,9 @@ import android.widget.Toast;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.example.pantrypal.Favorites;
+import com.example.pantrypal.NewRecipe;
+import com.example.pantrypal.R;
 import com.example.pantrypal.apiTools.MealDBRecipe;
 import com.example.pantrypal.apiTools.RecipeRetriever;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
