+++ app/src/main/java/com/example/pantrypal/activities/ViewRecipe.java
@@ -1,4 +1,4 @@
-package com.example.pantrypal;
+package com.example.pantrypal.activities;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -14,6 +14,11 @@ import android.widget.Toast;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.example.pantrypal.Comment;
+import com.example.pantrypal.CommentsListAdapter;
+import com.example.pantrypal.Favorites;
+import com.example.pantrypal.NewRecipe;
+import com.example.pantrypal.R;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
