+++ app/src/main/res/layout/activity_login.xml
@@ -6,7 +6,7 @@
     android:gravity="center"
     android:orientation="vertical"
     android:padding="15dp"
-    tools:context=".Login">
+    tools:context=".activities.Login">
 
 
     <TextView
