+++ app/src/main/res/layout/activity_registration.xml
@@ -7,7 +7,7 @@
     android:gravity="center"
     android:padding="15dp"
     android:layout_height="match_parent"
-    tools:context=".Registration">
+    tools:context=".activities.Registration">
 
     <TextView
         android:text="@string/Register"
