+++ app/src/main/res/layout/activity_search.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/yellow"
-    tools:context=".Search">
+    tools:context=".activities.Search">
 
     <ListView
         android:id="@+id/listView"
