+++ app/src/main/res/layout/activity_viewrecipe.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#F6C28A"
-    tools:context=".ViewRecipe">
+    tools:context=".activities.ViewRecipe">
 
     <ScrollView
         android:layout_width="match_parent"
